Only run postun when uninstall
postun is running for yum reinstall on rhel8, so that makes /usr/local/greenplum-db
removed after yum reinstall on rhel8. postun does not run for yum resintall on rhel6/7,
So inorder to fix the error, we make a if condition, only when $1=0 which means when
yum remove is called, the postun will be called.
Authored-by: Shaoqi Bai <bshaoqi@vmware.com>

+++ ci/concourse/scripts/gpdb-clients.spec
@@ -91,6 +91,8 @@ ln -sT $RPM_INSTALL_PREFIX/%{name}-%{gpdb_clients_version} $RPM_INSTALL_PREFIX/%
 sed -i -e "1 s~^\(GPHOME_CLIENTS=\).*~\1$RPM_INSTALL_PREFIX/%{name}-%{gpdb_clients_version}~" $RPM_INSTALL_PREFIX/%{name}-%{gpdb_clients_version}/greenplum_clients_path.sh
 
 %postun
-if [[ $(readlink $RPM_INSTALL_PREFIX/%{name}) == $RPM_INSTALL_PREFIX/%{name}-%{gpdb_clients_version} ]]; then
-  unlink $RPM_INSTALL_PREFIX/%{name}
+if [ $1 -eq 0 ] ; then
+  if [[ $(readlink $RPM_INSTALL_PREFIX/%{name}) == $RPM_INSTALL_PREFIX/%{name}-%{gpdb_clients_version} ]]; then
+    unlink $RPM_INSTALL_PREFIX/%{name}
+  fi
 fi
